+++ code/default/gae_proxy/local/ipv6_tunnel/__init__.py
@@ -6,7 +6,7 @@ from .common import test_teredo
 
 if "arm" in platform.machine():
     from .unknown import state, state_pp, switch_pp, enable, disable, set_best_server
-elif sys.platform == "win32" and platform.release() != "XP":
+elif sys.platform == "win32":
     from .win10 import state, state_pp, switch_pp, enable, disable, set_best_server
 elif sys.platform.startswith("linux"):
     from .linux import state, state_pp, switch_pp, enable, disable, set_best_server
